+++ dat257/src/css/Heatmap.css
@@ -1,0 +1,20 @@
+.flexbox-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: left;
+    gap: 1px;
+}
+
+.week-heatmap-size{
+    padding: 2em;
+    height: 100%;
+    width: 100%;
+}
+
+.year-heatmap-size{
+    padding: 2em;
+    height: 100%;
+    width: 250%;
+    overflow: auto;
+}
+
